+++ packages/assets-controllers/src/TokensController.test.ts
@@ -17,8 +17,10 @@ import type { InternalAccount } from '@metamask/keyring-api';
 import type {
   NetworkClientConfiguration,
   NetworkClientId,
+  NetworkState,
 } from '@metamask/network-controller';
 import { getDefaultNetworkControllerState } from '@metamask/network-controller';
+import type { Patch } from 'immer';
 import nock from 'nock';
 import * as sinon from 'sinon';
 import { v1 as uuidV1 } from 'uuid';
@@ -196,6 +198,84 @@ describe('TokensController', () => {
         name: undefined,
       });
     });
+  });
+
+  it('should add tokens and update existing ones and detected tokens', async () => {
+    const selectedAddress = '0x0001';
+    const selectedAccount = createMockInternalAccount({
+      address: selectedAddress,
+    });
+    await withController(
+      {
+        mockNetworkClientConfigurationsByNetworkClientId: {
+          networkClientId1: buildCustomNetworkClientConfiguration({
+            chainId: '0x1',
+          }),
+        },
+        mocks: {
+          getSelectedAccount: selectedAccount,
+          getAccount: selectedAccount,
+        },
+      },
+      async ({ controller }) => {
+        await controller.addDetectedTokens(
+          [
+            {
+              address: '0x01',
+              symbol: 'barA',
+              decimals: 2,
+            },
+          ],
+          {
+            selectedAddress: '0x0001',
+            chainId: '0x1',
+          },
+        );
+
+        await controller.addTokens(
+          [
+            {
+              address: '0x01',
+              symbol: 'barA',
+              decimals: 2,
+              aggregators: [],
+              name: 'Token1',
+            },
+            {
+              address: '0x02',
+              symbol: 'barB',
+              decimals: 2,
+              aggregators: [],
+              name: 'Token2',
+            },
+          ],
+          'networkClientId1',
+        );
+
+        expect(controller.state.allTokens).toStrictEqual({
+          '0x1': {
+            '0x0001': [
+              {
+                address: '0x01',
+                symbol: 'barA',
+                decimals: 2,
+                aggregators: [],
+                name: 'Token1',
+                image: undefined,
+              },
+              {
+                address: '0x02',
+                symbol: 'barB',
+                decimals: 2,
+                aggregators: [],
+                name: 'Token2',
+                image: undefined,
+              },
+            ],
+          },
+        });
+      },
+    );
   });
 
   it('should add detected tokens', async () => {
@@ -679,6 +759,86 @@ describe('TokensController', () => {
             [ChainId.goerli]: {
               [selectedAddress1]: ['0x02'],
               [selectedAddress2]: ['0x03'],
+            },
+          });
+        },
+      );
+    });
+
+    it('should ignore tokens by networkClientId', async () => {
+      const selectedAddress = '0x0001';
+      const otherAddress = '0x0002';
+      const selectedAccount = createMockInternalAccount({
+        address: selectedAddress,
+      });
+      const otherAccount = createMockInternalAccount({
+        address: otherAddress,
+      });
+
+      await withController(
+        {
+          mocks: {
+            getSelectedAccount: selectedAccount,
+            getAccount: selectedAccount,
+          },
+        },
+        async ({ controller, triggerSelectedAccountChange, changeNetwork }) => {
+          // Select the first account
+          triggerSelectedAccountChange(selectedAccount);
+
+          // Add and ignore a token on Sepolia
+          changeNetwork({ selectedNetworkClientId: InfuraNetworkType.sepolia });
+          await controller.addToken({
+            address: '0x01',
+            symbol: 'Token1',
+            decimals: 18,
+          });
+          expect(controller.state.tokens).toHaveLength(1);
+          expect(controller.state.ignoredTokens).toHaveLength(0);
+
+          controller.ignoreTokens(['0x01'], InfuraNetworkType.sepolia);
+          expect(controller.state.tokens).toHaveLength(0);
+          expect(controller.state.ignoredTokens).toStrictEqual(['0x01']);
+
+          // Verify that Goerli network has no ignored tokens
+          changeNetwork({ selectedNetworkClientId: InfuraNetworkType.goerli });
+          expect(controller.state.ignoredTokens).toHaveLength(0);
+
+          // Add and ignore a token on Goerli
+          await controller.addToken({
+            address: '0x02',
+            symbol: 'Token2',
+            decimals: 8,
+          });
+          controller.ignoreTokens(['0x02'], InfuraNetworkType.goerli);
+          expect(controller.state.tokens).toHaveLength(0);
+          expect(controller.state.ignoredTokens).toStrictEqual(['0x02']);
+
+          // Verify that switching back to Sepolia retains its ignored tokens
+          changeNetwork({ selectedNetworkClientId: InfuraNetworkType.sepolia });
+          expect(controller.state.ignoredTokens).toStrictEqual(['0x01']);
+
+          // Switch to a different account on Goerli
+          triggerSelectedAccountChange(otherAccount);
+          expect(controller.state.ignoredTokens).toHaveLength(0);
+
+          // Add and ignore a token on the new account
+          await controller.addToken({
+            address: '0x03',
+            symbol: 'Token3',
+            decimals: 6,
+          });
+          controller.ignoreTokens(['0x03'], InfuraNetworkType.goerli);
+          expect(controller.state.ignoredTokens).toStrictEqual(['0x03']);
+
+          // Validate the overall ignored tokens state
+          expect(controller.state.allIgnoredTokens).toStrictEqual({
+            [ChainId.sepolia]: {
+              [selectedAddress]: ['0x01'],
+            },
+            [ChainId.goerli]: {
+              [selectedAddress]: ['0x02'],
+              [otherAddress]: ['0x03'],
             },
           });
         },
@@ -2142,6 +2302,66 @@ describe('TokensController', () => {
         },
       );
     });
+
+    it('should clear allDetectedTokens under chain ID and selected address when a detected token is added to tokens list', async () => {
+      const selectedAddress = '0x1';
+      const selectedAccount = createMockInternalAccount({
+        address: selectedAddress,
+      });
+      const tokenAddress = '0x01';
+      const dummyDetectedTokens = [
+        {
+          address: tokenAddress,
+          symbol: 'barA',
+          decimals: 2,
+          aggregators: [],
+          isERC721: undefined,
+          name: undefined,
+          image: undefined,
+        },
+      ];
+      const dummyTokens = [
+        {
+          address: tokenAddress,
+          symbol: 'barA',
+          decimals: 2,
+          aggregators: [],
+          isERC721: undefined,
+          name: undefined,
+          image: undefined,
+        },
+      ];
+
+      await withController(
+        {
+          options: {
+            chainId: ChainId.mainnet,
+          },
+          mocks: {
+            getSelectedAccount: selectedAccount,
+          },
+        },
+        async ({ controller }) => {
+          // First, add detected tokens
+          await controller.addDetectedTokens(dummyDetectedTokens);
+          expect(
+            controller.state.allDetectedTokens[ChainId.mainnet][
+              selectedAddress
+            ],
+          ).toStrictEqual(dummyDetectedTokens);
+
+          // Now, add the same token to the tokens list
+          await controller.addTokens(dummyTokens);
+
+          // Check that allDetectedTokens for the selected address is cleared
+          expect(
+            controller.state.allDetectedTokens[ChainId.mainnet][
+              selectedAddress
+            ],
+          ).toStrictEqual([]);
+        },
+      );
+    });
   });
 
   describe('when TokenListController:stateChange is published', () => {
@@ -2309,6 +2529,64 @@ describe('TokensController', () => {
     });
   });
 
+  describe('when NetworkController:stateChange is published', () => {
+    it('removes tokens for removed networks', async () => {
+      const initialState = {
+        allTokens: {
+          '0x1': {
+            '0x134': [
+              {
+                address: '0x01',
+                symbol: 'TKN1',
+                decimals: 18,
+                aggregators: [],
+                name: 'Token 1',
+              },
+            ],
+          },
+          '0x5': {
+            // goerli
+            '0x456': [
+              {
+                address: '0x02',
+                symbol: 'TKN2',
+                decimals: 18,
+                aggregators: [],
+                name: 'Token 2',
+              },
+            ],
+          },
+        },
+        tokens: [],
+        ignoredTokens: [],
+        detectedTokens: [],
+        allIgnoredTokens: {},
+        allDetectedTokens: {},
+      };
+
+      await withController(
+        { options: { state: initialState } },
+        async ({ controller, triggerNetworkStateChange }) => {
+          // Verify initial state
+          expect(controller.state).toStrictEqual(initialState);
+
+          // Simulate removing goerli
+          triggerNetworkStateChange({} as NetworkState, [
+            {
+              op: 'remove',
+              path: ['networkConfigurationsByChainId', '0x5'],
+            },
+          ]);
+
+          // Verify tokens were removed on goerli
+          expect(controller.state.allTokens).toStrictEqual({
+            '0x1': initialState.allTokens['0x1'],
+          });
+        },
+      );
+    });
+  });
+
   describe('resetState', () => {
     it('resets the state to default state', async () => {
       const initialState: TokensControllerState = {
@@ -2406,6 +2684,10 @@ type WithControllerCallback<ReturnValue> = ({
   messenger: UnrestrictedMessenger;
   approvalController: ApprovalController;
   triggerSelectedAccountChange: (internalAccount: InternalAccount) => void;
+  triggerNetworkStateChange: (
+    networkState: NetworkState,
+    patches: Patch[],
+  ) => void;
   getAccountHandler: jest.Mock;
   getSelectedAccountHandler: jest.Mock;
 }) => Promise<ReturnValue> | ReturnValue;
@@ -2477,6 +2759,7 @@ async function withController<ReturnValue>(
     ],
     allowedEvents: [
       'NetworkController:networkDidChange',
+      'NetworkController:stateChange',
       'AccountsController:selectedEvmAccountChange',
       'TokenListController:stateChange',
     ],
@@ -2536,12 +2819,20 @@ async function withController<ReturnValue>(
     getNetworkClientById,
   );
 
+  const triggerNetworkStateChange = (
+    networkState: NetworkState,
+    patches: Patch[],
+  ) => {
+    messenger.publish('NetworkController:stateChange', networkState, patches);
+  };
+
   return await fn({
     controller,
     changeNetwork,
     messenger,
     approvalController,
     triggerSelectedAccountChange,
+    triggerNetworkStateChange,
     getAccountHandler,
     getSelectedAccountHandler,
   });